# MyBGList/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MyBGList Web API",
    "version": "1.0"
  },
  "paths": {
...
    "/auth/test/1": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Auth test #1 (authenticated users).",
        "description": "Returns 200 - OK if called by an authenticated user regardless of its role(s).",
        "responses": {
          "200": {
            "description": "Authorized"
          },
          "401": {
            "description": "Not authorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/auth/test/2": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Auth test #2 (Moderator role).",
        "description": "Returns 200 - OK status code if called by an authenticated user assigned to the Moderator role.",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/auth/test/3": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Auth test #3 (Administrator role).",
        "description": "Returns 200 - OK if called by an authenticated user assigned to the Administrator role.",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}